requirements.txt
fastapi==0.95.2
pydantic==1.10.13
starlette==0.27.0

uvicorn==0.30.0
psycopg2-binary==2.9.11

sqlalchemy==2.0.25
alembic==1.13.1

# avoids Rust build issues
python-dotenv==1.0.1

openai==0.28.1
faiss-cpu==1.13.2
